Rework HAKNXLocationsRepository constructors

+++ KNXFunctionAnalyzer/HAKNXLocationsRepository.py
@@ -20,15 +20,20 @@ class HAKNXLocationsRepository:
             return dumper.represent_scalar('tag:yaml.org,2002:str', value, style='"')
         return dumper.represent_data(value)
 
-    def __init__(self, knx_spaces_repository: KNXSpacesRepository, knx_project_manager: KNXProjectManager):
+    def __init__(self):
         self._locations_list = []
         # Add the custom representer to the PyYAML instance
         yaml.add_representer(Quoted, self.__custom_representer)
+
+    def import_from_knx_spaces_repository(self, knx_spaces_repository: KNXSpacesRepository, knx_project_manager: KNXProjectManager):
         for name, element in knx_spaces_repository:
             location = HAKNXLocation(element, knx_project_manager)
             location._name = name
             if not location.is_empty():
                 self.add_location(location)
+
+    def import_from_file(self):
+        pass
 
     def add_location(self, location: HAKNXLocation):
         self._locations_list.append(location)
@@ -43,16 +48,19 @@ class HAKNXLocationsRepository:
     def __next__(self):
         return self._locations_list.__iter__().__next__()
 
-    def dump(self, output_path, create_output_path):
+    def dump(self, output_path, create_output_path: bool = False, overwrite: bool = False):
         if not os.path.exists(output_path):
             if create_output_path:
                 os.makedirs(output_path, exist_ok=True)
             else:
-                raise ValueError(f"Output path '{output_path}' does not exist.")
+                raise FileNotFoundError(f"Output path '{output_path}' does not exist.")
         if not os.path.isdir(output_path):
-            raise ValueError(f"Output path '{output_path}' is not a directory.")
+            raise NotADirectoryError(f"Output path '{output_path}' is not a directory.")
         for element in self._locations_list:
             file_path = os.path.join(output_path, f"{element.get_name()}.yaml")
-            with open(file_path, "w") as file:
-                file.write(element.dump())
+            if os.path.exists(file_path) and not overwrite:
+                raise PermissionError(f"File '{file_path}' already exists.")
+            else:
+                with open(file_path, "w") as file:
+                    file.write(element.dump())
